point: Use f32 for coordinates

+++ src/point.rs
@@ -1,12 +1,12 @@
 #[derive(Clone)]
 pub struct Point {
-    pub x: i32,
-    pub y: i32,
-    pub z: i32
+    pub x: f32,
+    pub y: f32,
+    pub z: f32
 }
 
 impl Point {
-    pub fn new(x: i32, y: i32, z: i32) -> Point {
+    pub fn new(x: f32, y: f32, z: f32) -> Point {
         Point {
             x: x,
             y: y,
